+++ raytracer/CHANGELOG.md
@@ -7,13 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
+### Changed
 
-- Support for creating `RealSpec` instances from RefractiveIndex.info dispersion formulas 1 - 9.
+- `RefractiveIndexSpec` is now a trait which supports getting refractive index data from any generic materials database.
 
 ### Fixed
 
-- Fixed an import error in the `n` macro that required `RealSpec` to be imported inside a module to use it.
+- Fixed an import error in the `n` macro.
 
 ## [0.2.0] 2024-12-09